Updated UsedSpace to int64
To prevent errors when numbers exceed 32 bit capacity.

+++ protonmail/users.go
@@ -7,7 +7,7 @@ import (
 type User struct {
 	ID         string
 	Name       string
-	UsedSpace  int
+	UsedSpace  int64
 	Currency   string // e.g. EUR
 	Credit     int
 	MaxSpace   int64
